requirements.txt
streamlit
anthropic
crewai
llama-index
pinecone-client
PyGithub
python-dotenv
fastmcp
httpx
pinecone
voyageai
